♻️ Refactor: 회원 탈퇴 로직 Auth -> User 로 이동
