+++ src/main/java/com/likelion/backendplus4/talkpick/backend/user/application/port/in/UserServiceUseCase.java
@@ -27,4 +27,14 @@ public interface UserServiceUseCase {
 	 * @since 2025-05-16
 	 */
 	void updateMyInfo(User user);
+
+	/**
+	 * 회원 정보를 영구 삭제합니다.
+	 *
+	 * @param id 삭제할 회원의 고유 식별자
+	 * @author 박찬병
+	 * @since 2025-05-12
+	 * @modified 2025-05-12
+	 */
+	void deleteUser(Long id);
 }
